MAINT: stats.somersd: avoid integer overflow (#18137)
* Bugfix for somersd where an integer overflow could occur for cases that should be computable (Issue #18132). Modified the somersd function and added unit tests.

+++ scipy/stats/_hypotests.py
@@ -871,7 +871,8 @@ def somersd(x, y=None, alternative='two-sided'):
         table = x
     else:
         raise ValueError("x must be either a 1D or 2D array")
-    d, p = _somers_d(table, alternative)
+    # The table type is converted to a float to avoid an integer overflow
+    d, p = _somers_d(table.astype(float), alternative)
 
     # add alias for consistency with other correlation functions
     res = SomersDResult(d, p, table)
